ui/pyproject.toml
[tool.poetry]
name = "ottobot"
version = "0.1.0"
description = "LLM-powered chat application"
authors = ["Youcef Kharroubi"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
langchain = "^0.2.11"
langchain-ollama = "^0.1.0"
streamlit = "^1.37.0"
langchain-experimental = "^0.0.63"
pydantic = "^2.8.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
